Move SlackMessage to common
