+++ kifi-backend/modules/common/app/com/keepit/slack/models/SlackMessage.scala
@@ -1,9 +1,7 @@
 package com.keepit.slack.models
 
-import com.keepit.common.db.slick.DataBaseComponent
-import com.keepit.common.strings._
-import play.api.libs.json._
 import play.api.libs.functional.syntax._
+import play.api.libs.json._
 import com.kifi.macros.json
 
 @json case class SlackUserId(value: String)
@@ -23,38 +21,6 @@ import com.kifi.macros.json
 }
 
 @json case class SlackMessageType(value: String)
-
-object SlackDbColumnTypes {
-  def userId(db: DataBaseComponent) = {
-    import db.Driver.simple._
-    MappedColumnType.base[SlackUserId, String](_.value, SlackUserId(_))
-  }
-  def username(db: DataBaseComponent) = {
-    import db.Driver.simple._
-    MappedColumnType.base[SlackUsername, String](_.value, SlackUsername(_))
-  }
-  def teamId(db: DataBaseComponent) = {
-    import db.Driver.simple._
-    MappedColumnType.base[SlackTeamId, String](_.value, SlackTeamId(_))
-  }
-  def teamName(db: DataBaseComponent) = {
-    import db.Driver.simple._
-    MappedColumnType.base[SlackTeamName, String](_.value, SlackTeamName(_))
-  }
-  def channelId(db: DataBaseComponent) = {
-    import db.Driver.simple._
-    MappedColumnType.base[SlackChannelId, String](_.value, SlackChannelId(_))
-  }
-  def channel(db: DataBaseComponent) = {
-    import db.Driver.simple._
-    MappedColumnType.base[SlackChannelName, String](_.value, SlackChannelName(_))
-  }
-
-  def timestamp(db: DataBaseComponent) = {
-    import db.Driver.simple._
-    MappedColumnType.base[SlackMessageTimestamp, String](_.value, SlackMessageTimestamp(_))
-  }
-}
 
 case class SlackAttachment(
   fallback: String,
@@ -136,44 +102,6 @@ object SlackAttachment {
     (__ \ "image_url").formatNullable[String] and
     (__ \ "image_thumb").formatNullable[String]
   )(applyFromSlack, unlift(unapplyToSlack))
-}
-
-case class OutgoingSlackMessage( // https://api.slack.com/incoming-webhooks
-  text: String,
-  channel: Option[SlackChannelName],
-  username: String,
-  iconUrl: String,
-  attachments: Seq[SlackAttachment],
-  unfurlLinks: Boolean,
-  unfurlMedia: Boolean)
-
-object OutgoingSlackMessage {
-
-  val kifiIconUrl = "https://d1dwdv9wd966qu.cloudfront.net/img/favicon64x64.7cc6dd4.png"
-
-  def fromKifi(text: String, channel: Option[SlackChannelName] = None, attachments: Seq[SlackAttachment] = Seq.empty) = OutgoingSlackMessage(
-    text,
-    channel = channel,
-    username = "Kifi",
-    iconUrl = kifiIconUrl,
-    attachments = attachments,
-    unfurlLinks = true,
-    unfurlMedia = true
-  )
-
-  implicit val writes: Writes[OutgoingSlackMessage] = Writes { o =>
-    Json.obj(
-      "text" -> o.text,
-      "channel" -> o.channel,
-      "username" -> o.username,
-      "icon_url" -> o.iconUrl,
-      "attachments" -> o.attachments,
-      "unfurl_links" -> o.unfurlLinks,
-      "unfurl_media" -> o.unfurlMedia
-    )
-  }
-
-  def escapeSegment(segment: String): String = segment.replaceAllLiterally("<" -> "&lt;", ">" -> "&gt;", "&" -> "&amp")
 }
 
 case class SlackMessage(
