+++ kifi-backend/modules/shoebox/app/com/keepit/payments/DescriptionElements.scala
@@ -3,7 +3,7 @@ package com.keepit.payments
 import com.keepit.common.mail.EmailAddress
 import com.keepit.common.path.Path
 import com.keepit.model.{ OrganizationRole, OrganizationHandle, BasicOrganization }
-import com.keepit.slack.models.{ OutgoingSlackMessage, SlackMessage }
+import com.keepit.slack.models.{ SlackMessageRequest, SlackMessage }
 import com.keepit.social.BasicUser
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
@@ -88,7 +88,7 @@ object DescriptionElements {
   def formatForSlack(description: DescriptionElements): String = {
     interpolatePunctuation(description.flatten).map { be =>
       be.url
-        .map(u => s"<$u|${OutgoingSlackMessage.escapeSegment(be.text)}>")
+        .map(u => s"<$u|${SlackMessageRequest.escapeSegment(be.text)}>")
         .getOrElse(be.text)
     }.mkString
   }
