+++ kifi-backend/modules/shoebox/app/com/keepit/slack/FakeSlackClientImpl.scala
@@ -6,11 +6,11 @@ import scala.collection.mutable
 import scala.concurrent.Future
 
 class FakeSlackClientImpl extends SlackClient {
-  val messagesByWebhook: mutable.Map[String, List[OutgoingSlackMessage]] = mutable.Map.empty.withDefaultValue(List.empty)
+  val messagesByWebhook: mutable.Map[String, List[SlackMessageRequest]] = mutable.Map.empty.withDefaultValue(List.empty)
 
   def identifyUser(token: SlackAccessToken): Future[SlackIdentifyResponse] = ???
   def processAuthorizationResponse(code: SlackAuthorizationCode): Future[SlackAuthorizationResponse] = ???
-  def sendToSlack(url: String, msg: OutgoingSlackMessage): Future[Unit] = {
+  def sendToSlack(url: String, msg: SlackMessageRequest): Future[Unit] = {
     messagesByWebhook.put(url, msg :: messagesByWebhook(url))
     Future.successful(())
   }
